SHA256 from OpenSSL
Changes to be committed:
	modified:   DataWallEngine/DataWallEngine.cpp
	modified:   DataWallEngine/DataWallEngine.h
	modified:   DataWallEngine/framework.h
	modified:   TestDataWall/TestDataWall.cpp

+++ DataWallEngine/DataWallEngine.h
@@ -51,7 +51,7 @@ namespace DataWallEngine
 
     extern "C" DATAWALLENGINE_API HRESULT EncryptData(BYTE* data, INT32 size, BYTE* key);
     extern "C" DATAWALLENGINE_API HRESULT DecryptData(BYTE * data, INT32 size, BYTE * key);
-    extern "C" DATAWALLENGINE_API HRESULT CalculateHash(BYTE * data, UINT64 size, UINT32* state);
+    extern "C" DATAWALLENGINE_API HRESULT CalculateHash(BYTE * data, UINT64 size, BYTE* state);
     //extern "C" DATAWALLENGINE_API HRESULT NetworkAuthentication();
 
     extern "C" DATAWALLENGINE_API HRESULT PackInContainer(BYTE* data, INT32 size, ContentType type, BYTE* key, const char* container_name);
